refactor(Domain): 당첨로또에서만 사용하는 클래스를 하나의 하위패키지로 묶어서 관리하도록 변경
